Case stage update script
